Update titles, config
Having the titles match the same text in the sidebar enables the blue marker
that indicates the current section of the document, which is a 508 feature.

Removed the `baseurl:` attribute from _config.yml, as that isn't needed to
publish on 18F Pages anymore. Consequently, the site can be viewed at
localhost:4000 rather than localhost:4000/agile/.